+++ ArenaService.Worker/Utils/LocalRandom.cs
@@ -1,0 +1,14 @@
+using Libplanet.Action;
+
+namespace ArenaService.Worker.Utils;
+
+public class LocalRandom : Random, IRandom
+{
+    public int Seed { get; }
+
+    public LocalRandom(int seed)
+        : base(seed)
+    {
+        Seed = seed;
+    }
+}
