+++ ArenaService.Worker/Worker.cs
@@ -1,16 +1,22 @@
-using ArenaService.Worker.Rpc;
-
 namespace ArenaService.Worker;
+
+using ArenaService.Worker.Rpc;
+using ArenaService.Worker.Utils;
+using Lib9c.Renderers;
+using Libplanet.Action;
+using Nekoyume.Action;
 
 public class Worker : BackgroundService
 {
     private readonly ILogger<Worker> _logger;
-    private readonly RpcClient _rpcClient;
+    private readonly ActionRenderer _actionRenderer;
 
-    public Worker(ILogger<Worker> logger, RpcClient client)
+    public Worker(ILogger<Worker> logger, ActionRenderer actionRenderer)
     {
         _logger = logger;
-        _rpcClient = client;
+
+        _actionRenderer = actionRenderer;
+        _actionRenderer.ActionRenderSubject.Subscribe(RenderAction);
     }
 
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
@@ -18,7 +24,26 @@ public class Worker : BackgroundService
         while (true)
         {
             _logger.LogInformation("worker");
-            await Task.Delay(1000, stoppingToken);
+            await Task.Delay(100000, stoppingToken);
+        }
+    }
+
+    public async void RenderAction(ActionEvaluation<ActionBase> ev)
+    {
+        if (ev.Exception is null)
+        {
+            var seed = ev.RandomSeed;
+            var random = new LocalRandom(seed);
+            var stateRootHash = ev.OutputState;
+            var hashBytes = stateRootHash.ToByteArray();
+            switch (ev.Action)
+            {
+                // Insert new product
+                case DailyReward d:
+                {
+                    break;
+                }
+            }
         }
     }
 }
