+++ ArenaService.Worker/appsettings.json
@@ -4,5 +4,8 @@
       "Default": "Information",
       "Microsoft.Hosting.Lifetime": "Information"
     }
+  },
+  "RpcConfig": {
+    "Port": 31238
   }
 }
